update project to latest

+++ app/src/main/java/com/example/smartwaterqualitymonitoringsystem/DashboardScreen.java
@@ -20,6 +20,7 @@ import com.google.android.gms.tasks.Task;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 
+
 public class DashboardScreen extends AppCompatActivity implements View.OnClickListener {
 
     CardView cardTemperature,cardSolids,cardTubidity,cardPHSensor,tv_history,cardWater,card_delete;
@@ -68,66 +69,46 @@ public class DashboardScreen extends AppCompatActivity implements View.OnClickLi
     {
         int id = view.getId();
 
-        switch (id)
-        {
-            case R.id.cardTemperature:
 
-                startActivity(new Intent(DashboardScreen.this,TemperatureScreen.class));
-                break;
-            case R.id.cardSolids:
-                startActivity(new Intent(DashboardScreen.this,SolidsScreen.class));
-                break;
+        if(id == R.id.cardTemperature){
+            startActivity(new Intent(DashboardScreen.this,TemperatureScreen.class));
+        }else if(id == R.id.cardSolids){
+            startActivity(new Intent(DashboardScreen.this,SolidsScreen.class));
+        }else if(id == R.id.cardTubidity){
+            startActivity(new Intent(DashboardScreen.this,TurbidityScreen.class));
+        }else if(id == R.id.cardPHSensor){
+            startActivity(new Intent(DashboardScreen.this,PHSensorScreen.class));
+        }else if(id == R.id.cardWater){
+            startActivity(new Intent(DashboardScreen.this,WaterLevelScreen.class));
+        }else if(id == R.id.tv_history){
+            startActivity(new Intent(DashboardScreen.this,ShowAllDataScreen.class));
+        }else if(id == R.id.card_delete){
+            deleteData();
+        }else if(id == R.id.img_logout){
+            AlertDialog.Builder builder = new AlertDialog.Builder(this);
+            builder.setCancelable(false);
+            builder.setTitle("Logout");
+            builder.setMessage("Are you sure to logout");
 
-            case R.id.cardTubidity:
-                startActivity(new Intent(DashboardScreen.this,TurbidityScreen.class));
-                break;
+            builder.setPositiveButton("Yes", new DialogInterface.OnClickListener() {
+                @Override
+                public void onClick(DialogInterface dialogInterface, int i) {
 
-            case R.id.cardPHSensor:
-                startActivity(new Intent(DashboardScreen.this,PHSensorScreen.class));
-                break;
+                    startActivity(new Intent(DashboardScreen.this,LoginScreen.class));
+                    finish();
 
-            case R.id.cardWater:
-                startActivity(new Intent(DashboardScreen.this,WaterLevelScreen.class));
-                break;
+                }
+            }).setNegativeButton("No", new DialogInterface.OnClickListener() {
+                @Override
+                public void onClick(DialogInterface dialogInterface, int i) {
 
-            case R.id.tv_history:
-                startActivity(new Intent(DashboardScreen.this,ShowAllDataScreen.class));
-                break;
+                }
+            });
 
-            case R.id.card_delete:
-                deleteData();
-                break;
-
-
-            case R.id.img_logout:
-                AlertDialog.Builder builder = new AlertDialog.Builder(this);
-                builder.setCancelable(false);
-                builder.setTitle("Logout");
-                builder.setMessage("Are you sure to logout");
-
-                builder.setPositiveButton("Yes", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialogInterface, int i) {
-
-                        startActivity(new Intent(DashboardScreen.this,LoginScreen.class));
-                        finish();
-
-                    }
-                }).setNegativeButton("No", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialogInterface, int i) {
-
-                    }
-                });
-
-                AlertDialog obj = builder.create();
-                obj.show();
-                break;
-
-
-
-
+            AlertDialog obj = builder.create();
+            obj.show();
         }
+
 
     }
